fix identifiers not being able to start with "d"
Prior to this the lexer would see the first "d" and return token.D.
"Dice" would be two tokens, "D" and "ice".

The solution I chose is to restrict the grammar so that:

1. You cannot have a space between "d" and the face count.
2. token.Word cannot follow the same format as token.D.

This way, the code now treats both as words initially before
checking if the word is actually a d-token. If it's a d-token, it
will change the token kind appropriately.

Integer conversions still work great because they just ignore any
non-digit, in this case the "d".